New scenes
Renamed old scenes and added new scene. NOTE: I am currently ignoring all textures, images and lightmaps because they are too heavy.
