+++ UnityDroneSim/Assets/ML-Agents.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b5b345839a314f4b9519e8ab3b1ce4b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
