+++ UnityDroneSim/Assets/Materials/SampleWorld.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c31ff5921f46445bf8cec897a818b626
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
