+++ UnityDroneSim/Assets/Materials/SampleWorld/Bridge.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 398868a3cafb24400801105e655388d4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
