+++ UnityDroneSim/Assets/Materials/SampleWorld/NewLayer 1.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cae444b62e2f34915a372d978ecb837a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8574412962073106934
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
