+++ UnityDroneSim/Assets/Materials/SampleWorld/NewLayer 2.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f3895798f754e4722ad0b6ecc5a4b4f1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8574412962073106934
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
