+++ UnityDroneSim/Assets/Materials/SampleWorld/NewLayer.terrainlayer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 173aab4527a0949eb910120585b3e6b1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8574412962073106934
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
