+++ UnityDroneSim/Assets/Materials/SampleWorld/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f8d40138dd5c04933ab37b9d23fd560c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
