+++ UnityDroneSim/Assets/Materials/SampleWorld/TowerBalcony.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7fa678da85e0943aa998128c2a1df8b2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
