+++ UnityDroneSim/Assets/Materials/SampleWorld/TowerDome.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6517aa4649970450d8774494699b7ba0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
