+++ UnityDroneSim/Assets/Materials/SampleWorld/TowerWall.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 01613a9fa4a8d4d7bb11e5f715dfcfe6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
