+++ UnityDroneSim/Assets/Materials/SampleWorld/WorldEdge.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19f43688ece5143528c17cd848279f3f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
