+++ UnityDroneSim/Assets/Prefabs/Drones/DroneAI_2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b9952ca56cfa34f57b164c79b53d6d73
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
