+++ UnityDroneSim/Assets/Rendering/DepthMap.shadergraph
@@ -294,10 +294,10 @@
         "m_Expanded": true,
         "m_Position": {
             "serializedVersion": "2",
-            "x": -283.5000915527344,
-            "y": 158.0,
-            "width": 145.00003051757813,
-            "height": 111.5
+            "x": -187.00001525878907,
+            "y": 116.0,
+            "width": 145.00001525878907,
+            "height": 111.50003051757813
         }
     },
     "m_Slots": [
