+++ UnityDroneSim/Assets/Rendering/Renderer.asset.meta
