+++ UnityDroneSim/Assets/Rendering/Renderer_Depth.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 840765f50c1264831b2c75cd75edeae7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
