+++ UnityDroneSim/Assets/Scenes/ValleyWorld.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e0d3a962bb3248fca151a51a42cd84e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
