--- UnityDroneSim/Assets/Scripts/Utils/Painter2DExtensions.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 67df9fad6bde4058ba3d234de8e12d24
-timeCreated: 1740316697
